Remove setup package usage from test helpers
