+++ e2e/helper/submit/hedera.go
@@ -21,23 +21,22 @@ import (
 	"time"
 
 	"github.com/hashgraph/hedera-sdk-go/v2"
-	"github.com/limechain/hedera-eth-bridge-validator/e2e/setup"
 )
 
-func HbarToBridgeAccount(setup *setup.Setup, memo string, amount int64) (*hedera.TransactionResponse, error) {
+func HbarToBridgeAccount(hederaClient *hedera.Client, bridgeAccount hedera.AccountID, memo string, amount int64) (*hedera.TransactionResponse, error) {
 	hbarSendAmount := hedera.HbarFromTinybar(amount)
 	hbarRemovalAmount := hedera.HbarFromTinybar(-amount)
 	fmt.Println(fmt.Sprintf("Sending [%v] Hbars through the Bridge. Transaction Memo: [%s]", hbarSendAmount, memo))
 
 	res, err := hedera.NewTransferTransaction().
-		AddHbarTransfer(setup.Clients.Hedera.GetOperatorAccountID(), hbarRemovalAmount).
-		AddHbarTransfer(setup.BridgeAccount, hbarSendAmount).
+		AddHbarTransfer(hederaClient.GetOperatorAccountID(), hbarRemovalAmount).
+		AddHbarTransfer(bridgeAccount, hbarSendAmount).
 		SetTransactionMemo(memo).
-		Execute(setup.Clients.Hedera)
+		Execute(hederaClient)
 	if err != nil {
 		return nil, err
 	}
-	rec, err := res.GetReceipt(setup.Clients.Hedera)
+	rec, err := res.GetReceipt(hederaClient)
 	if err != nil {
 		return nil, err
 	}
@@ -48,18 +47,18 @@ func HbarToBridgeAccount(setup *setup.Setup, memo string, amount int64) (*hedera
 	return &res, err
 }
 
-func TokensToBridgeAccount(setup *setup.Setup, tokenID hedera.TokenID, memo string, amount int64) (*hedera.TransactionResponse, error) {
+func TokensToBridgeAccount(hederaClient *hedera.Client, bridgeAccount hedera.AccountID, tokenID hedera.TokenID, memo string, amount int64) (*hedera.TransactionResponse, error) {
 	fmt.Println(fmt.Sprintf("Sending [%v] Tokens to the Bridge. Transaction Memo: [%s]", amount, memo))
 
 	res, err := hedera.NewTransferTransaction().
 		SetTransactionMemo(memo).
-		AddTokenTransfer(tokenID, setup.Clients.Hedera.GetOperatorAccountID(), -amount).
-		AddTokenTransfer(tokenID, setup.BridgeAccount, amount).
-		Execute(setup.Clients.Hedera)
+		AddTokenTransfer(tokenID, hederaClient.GetOperatorAccountID(), -amount).
+		AddTokenTransfer(tokenID, bridgeAccount, amount).
+		Execute(hederaClient)
 	if err != nil {
 		return nil, err
 	}
-	rec, err := res.GetReceipt(setup.Clients.Hedera)
+	rec, err := res.GetReceipt(hederaClient)
 	if err != nil {
 		return nil, err
 	}
@@ -70,7 +69,7 @@ func TokensToBridgeAccount(setup *setup.Setup, tokenID hedera.TokenID, memo stri
 	return &res, err
 }
 
-func NFTWithFeeToBridgeAccount(setup *setup.Setup, memo string, token string, serialNum int64, fee int64) (*hedera.TransactionResponse, error) {
+func NFTWithFeeToBridgeAccount(hederaClient *hedera.Client, bridgeAccount hedera.AccountID, memo string, token string, serialNum int64, fee int64) (*hedera.TransactionResponse, error) {
 	hbarSendAmount := hedera.HbarFromTinybar(fee)
 	hbarRemovalAmount := hedera.HbarFromTinybar(-fee)
 
@@ -81,15 +80,15 @@ func NFTWithFeeToBridgeAccount(setup *setup.Setup, memo string, token string, se
 	}
 
 	res, err := hedera.NewTransferTransaction().
-		AddNftTransfer(nftID, setup.Clients.Hedera.GetOperatorAccountID(), setup.BridgeAccount).
-		AddHbarTransfer(setup.Clients.Hedera.GetOperatorAccountID(), hbarRemovalAmount).
-		AddHbarTransfer(setup.BridgeAccount, hbarSendAmount).
+		AddNftTransfer(nftID, hederaClient.GetOperatorAccountID(), bridgeAccount).
+		AddHbarTransfer(hederaClient.GetOperatorAccountID(), hbarRemovalAmount).
+		AddHbarTransfer(bridgeAccount, hbarSendAmount).
 		SetTransactionMemo(memo).
-		Execute(setup.Clients.Hedera)
+		Execute(hederaClient)
 	if err != nil {
 		return nil, err
 	}
-	rec, err := res.GetReceipt(setup.Clients.Hedera)
+	rec, err := res.GetReceipt(hederaClient)
 	if err != nil {
 		return nil, err
 	}
